amqp: Log warning on unexpected application data bag
When we receive a relation change event with non-existing
expected keys in the application data bag, chances are the charm
in the other end of the relation is incompatible.  Log this fact.

Be more verbose when debug logging is enabled and log the contents
of the relation data in AMQPProvides event handlers.