Wait for unit to have bind address for peers
Rabbit wont start up on recent versions of juju as the network
information is delayed until the relation is established in order to
provide the correct bind address information for the relation. The charm
uses the amqp relation for recording information into the
rabbitmq-env.conf file, but this should be the peer address as the
rabbit internal logic uses it to determine if it has seen this node or
not.

Change the address recorded in rabbitmq-env.conf from the amqp binding
address to the peers binding address. Additionally, wait until the peers
binding address exists before trying to process remote relations and
handle a 401 Unauthorized error from rabbit when the node is started but
the unit has yet to configure the operator user.

Signed-off-by: Billy Olsen <billy.olsen@gmail.com>